+++ packages/react-network/src/context.ts
@@ -5,8 +5,13 @@ import {NetworkManager} from './manager';
 
 export const NetworkContext = React.createContext<NetworkManager | null>(null);
 
-export const HeaderContext = React.createContext<Record<string, string>>({});
-export const HeaderUniversalProvider = createUniversalProvider(
-  'request-headers',
-  HeaderContext,
+export interface UniversalNetworkDetails {
+  headers: Record<string, string>;
+}
+export const UniversalContext = React.createContext<UniversalNetworkDetails | null>(
+  null,
+);
+export const UniversalProvider = createUniversalProvider(
+  'network-details',
+  UniversalContext,
 );